+++ test/LinearLeastSquaresTests.jl
@@ -27,6 +27,8 @@ end
     n = length(fns)
     X = -1.0:0.125:2.0
     eqs = @inferred NormalEquations{n,Float64}()
+    @test eqs isa NormalEquations{n,Float64}
+    @test @inferred(eltype(eqs)) === Float64
     A = @inferred lhs_matrix(eqs; readonly=false)
     b = @inferred rhs_vector(eqs; readonly=false)
     @test size(A) == (n, n)
@@ -40,13 +42,36 @@ end
     end
     c = @inferred solve(eqs)
     @test collect(c) ≈ collect(c0)
+    # idem with weights
+    eqs1 = @inferred zero(eqs)
+    @test eqs1 === @inferred NormalEquations{n,eltype(eqs)}()
+    @test eqs1 === @inferred NormalEquations{n,eltype(eqs),(n*(n+1))>>1}()
+    @test_throws Exception NormalEquations{n,Int16}() # T must be floating-point
+    @test_throws Exception NormalEquations{n,Float32,1}() # invalid L
+    eqs2 = eqs1
+    w = 1.0
+    for x in X
+        fx = mcall(fns, x)
+        y = mdot(c0, fx)
+        w *= 1.1
+        eqs1 = @inferred update(eqs1, w, y, fx)
+        eqs2 = @inferred update(eqs2, y, fx...; wgt=w)
+    end
+    @test collect(eqs1.A) ≈ collect(eqs2.A)
+    @test collect(eqs1.b) ≈ collect(eqs2.b)
+    c1 = @inferred solve(eqs1)
+    c2 = @inferred solve(eqs2)
+    @test collect(c1) ≈ collect(c0)
+    @test collect(c2) ≈ collect(c0)
 
     # Fit a straight line plus sine.
-    fns = (x -> 1.0, identity, sin)
+    fns = (x -> true, identity, sin) # `true` is to force a conversion
     c0 = (-7.25, -2.50, 3.75)
     n = length(fns)
     X = -1.0:0.125:2.0
-    eqs = @inferred NormalEquations{n,Float64}()
+    eqs = @inferred NormalEquations{n,Float32}()
+    @test eqs isa NormalEquations{n,Float32}
+    @test @inferred(eltype(eqs)) === Float32
     A = @inferred lhs_matrix(eqs; readonly=false)
     b = @inferred rhs_vector(eqs; readonly=false)
     @test size(A) == (n, n)